Add Violin stats element
