+++ holoviews/element/stats.py
@@ -122,3 +122,30 @@ class Distribution(StatisticsElement):
     # Ensure Interface does not add an index
     _auto_indexable_1d = False
 
+
+class BoxWhisker(Chart):
+    """
+    BoxWhisker represent data as a distributions highlighting the
+    median, mean and various percentiles. It may have a single value
+    dimension and any number of key dimensions declaring the grouping
+    of each violin.
+    """
+
+    group = param.String(default='BoxWhisker', constant=True)
+
+    kdims = param.List(default=[], bounds=(0,None))
+
+    vdims = param.List(default=[Dimension('y')], bounds=(1,1))
+
+    _auto_indexable_1d = False
+
+
+class Violin(BoxWhisker):
+    """
+    Violin elements represent data as 1D distributions visualized
+    as a kernel-density estimate. It may have a single value dimension
+    and any number of key dimensions declaring the grouping of each
+    violin.
+    """
+
+    group = param.String(default='Violin')
