+++ holoviews/plotting/mpl/__init__.py
@@ -175,7 +175,9 @@ Store.register({Curve: CurvePlot,
 
                 # Statistics elements
                 Distribution: DistributionPlot,
-                Bivariate: BivariatePlot}, 'matplotlib', style_aliases=style_aliases)
+                Bivariate: BivariatePlot,
+                Violin: ViolinPlot},
+               'matplotlib', style_aliases=style_aliases)
 
 
 MPLPlot.sideplots.update({Histogram: SideHistogramPlot,
@@ -279,3 +281,5 @@ options.EdgePaths = Options('style', color='black')
 # Statistics
 options.Distribution = Options('style', facecolor=Cycle(), edgecolor='black',
                                alpha=0.5)
+options.Violin = Options('style', facecolors=Cycle(), bw_method='scott',
+                         showmedians=True, showextrema=False, alpha=0.7)
