+++ holoviews/plotting/mpl/chart.py
@@ -1012,78 +1012,3 @@ class SideSpikesPlot(AdjoinedPlot, SpikesPlot):
         all axis labels including ticks and ylabel. Valid options are 'left',
         'right', 'bare' 'left-bare' and 'right-bare'.""")
 
-
-
-class BoxPlot(ChartPlot):
-    """
-    BoxPlot plots the ErrorBar Element type and supporting
-    both horizontal and vertical error bars via the 'horizontal'
-    plot option.
-    """
-
-    style_opts = ['notch', 'sym', 'whis', 'bootstrap',
-                  'conf_intervals', 'widths', 'showmeans',
-                  'show_caps', 'showfliers', 'boxprops',
-                  'whiskerprops', 'capprops', 'flierprops',
-                  'medianprops', 'meanprops', 'meanline']
-
-    _plot_methods = dict(single='boxplot')
-
-    def get_extents(self, element, ranges):
-        return (np.NaN,)*4
-
-
-    def get_data(self, element, ranges, style):
-        groups = element.groupby(element.kdims)
-
-        data, labels = [], []
-
-        groups = groups.data.items() if element.kdims else [(element.label, element)]
-        for key, group in groups:
-            if element.kdims:
-                label = ','.join([d.pprint_value(v) for d, v in zip(element.kdims, key)])
-            else:
-                label = key
-            data.append(group[group.vdims[0]])
-            labels.append(label)
-        style['labels'] = labels
-        style = {k: v for k, v in style.items()
-                 if k not in ['zorder', 'label']}
-        style['vert'] = not self.invert_axes
-        format_kdims = [kd(value_format=None) for kd in element.kdims]
-        return (data,), style, {'dimensions': [format_kdims,
-                                               element.vdims[0]]}
-
-
-    def teardown_handles(self):
-        for group in self.handles['artist'].values():
-            for v in group:
-                v.remove()
-
-
-class SideBoxPlot(AdjoinedPlot, BoxPlot):
-
-    bgcolor = param.Parameter(default=(1, 1, 1, 0), doc="""
-        Make plot background invisible.""")
-
-    border_size = param.Number(default=0, doc="""
-        The size of the border expressed as a fraction of the main plot.""")
-
-    xaxis = param.ObjectSelector(default='bare',
-                                 objects=['top', 'bottom', 'bare', 'top-bare',
-                                          'bottom-bare', None], doc="""
-        Whether and where to display the xaxis, bare options allow suppressing
-        all axis labels including ticks and xlabel. Valid options are 'top',
-        'bottom', 'bare', 'top-bare' and 'bottom-bare'.""")
-
-    yaxis = param.ObjectSelector(default='bare',
-                                 objects=['left', 'right', 'bare', 'left-bare',
-                                          'right-bare', None], doc="""
-        Whether and where to display the yaxis, bare options allow suppressing
-        all axis labels including ticks and ylabel. Valid options are 'left',
-        'right', 'bare' 'left-bare' and 'right-bare'.""")
-
-    def __init__(self, *args, **kwargs):
-        super(SideBoxPlot, self).__init__(*args, **kwargs)
-        if self.adjoined:
-            self.invert_axes = not self.invert_axes
